src: entrances: Merge the entrance, exit times into the accessTime
It is so ambiguous to distinguish if someone enters into a place or not.
And, this makes development very hard to handle the exceptions such as
the malfunctioning of a BLE detecting device. So it is better that
merging those two time columns into one and think about this later.

And apply ESLint-ing.

Signed-off-by: Deokgyu Yang <secugyu@gmail.com>